requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Excel handling
openpyxl>=3.0.0
xlrd>=2.0.0

# PDF processing
PyPDF2>=3.0.0

# Config
pyyaml>=6.0

# HTTP (update checks, optional)
requests>=2.31.0

# Windows-specific (Outlook COM integration)
pywin32>=305 ; platform_system=="Windows"
